Update sketch.py
Adding a lot of comments.
I'm sure I'm missing some of the actual meat of the change, but the main thing was profiling and figuring out what is slow.
It appears that pandas dataframe is a poor data structure for this workload. 
A simple array with some helper indexer functions might be faster. 

There's more orchestration functionality that I want to eventually bring in, but for now I am focusing on the slow calculation issue.